+++ src/mwax_mover/queues/queue_worker.py
@@ -15,6 +15,7 @@ import time
 
 from mwax_mover import constants
 from mwax_mover.core.command import run_command
+from mwax_mover.core.timing import interruptible_sleep, sleep
 
 logger = logging.getLogger(__name__)
 
@@ -147,7 +148,7 @@ class QueueWorker:
         while self._running:
             if self._paused:
                 # if paused, put in a sleep to slow the wheel spinning
-                time.sleep(0.1)
+                sleep(0.1)
             else:
                 try:
                     success = False
@@ -208,7 +209,7 @@ class QueueWorker:
                                 " failures. Backing off for"
                                 f" {backoff} seconds."
                             )
-                            self.event.wait(backoff)
+                            interruptible_sleep(self.event, backoff)
 
                             # If this option is set, add item back to the end of
                             # the queue